Small fix, forgot to update the memory calculation for InterfaceFlags when switched from uint8_t to uint16_t

+++ src/muscl_block/SolverHydroMusclBlock.cpp
@@ -136,7 +136,7 @@ SolverHydroMusclBlock::SolverHydroMusclBlock(HydroParams& params,
 
   // flags data array for faces on 2:1 borders
   Interface_flags = FlagArrayBlock("Flags", nbOctsPerGroup);
-  total_mem_size += nbOctsPerGroup*sizeof(uint8_t);
+  total_mem_size += nbOctsPerGroup*sizeof(uint16_t);
 
   // all intermediate data array are sized upon nbOctsPerGroup
